refactor pages/index.tsx - the dashboard
